refactor: modularize app structure
