+++ index.html
@@ -65,6 +65,7 @@
       </div>
     </div>
 
-    <script type="module" src="layout.js"></script>
+    <script type="module" src="app.js"></script>
+    <script type="module" src="src/input/handlers.js"></script>
   </body>
 </html>
