+++ src/core/constants.js
@@ -1,0 +1,7 @@
+// constants.js
+// Defines tunable constants for the StarWiki visualization.
+export const R_MIN = 8;
+export const R_MAX = 40;
+export const SEGMENT_DIST = 40;
+export const MAX_GHOSTS = 5;
+export const RETURN_COLOR = 0xf7768e;
