+++ src/core/math.js
@@ -1,0 +1,17 @@
+// math.js
+// Miscellaneous math helpers including hashing and direction derivation.
+export function seededHash(str){
+  let h = 2166136261 >>> 0;
+  for (let i=0;i<str.length;i++){
+    h ^= str.charCodeAt(i);
+    h = Math.imul(h, 16777619);
+  }
+  return h >>> 0;
+}
+
+export function directionFromTitle(title){
+  const h = seededHash(title);
+  const theta = (h % 360) * Math.PI/180;
+  const phi = ((h>>9)%360) * Math.PI/180;
+  return [Math.cos(theta)*Math.sin(phi), Math.sin(theta)*Math.sin(phi), Math.cos(phi)];
+}
