+++ src/graph/layout.js
@@ -1,0 +1,15 @@
+// layout.js
+// Geometry helpers for positioning and visibility of neighbor stars.
+import { R_MIN, R_MAX } from '../core/constants.js';
+import { directionFromTitle } from '../core/math.js';
+
+export function positionForNeighbor(title, index, total){
+  const dir = directionFromTitle(title);
+  const r = R_MIN + (total <= 1 ? 0 : index/(total-1)) * (R_MAX - R_MIN);
+  return [dir[0]*r, dir[1]*r, dir[2]*r];
+}
+
+export function opacityFromRank(rank, total){
+  const t = total <= 1 ? 0 : rank/(total-1);
+  return 0.25 + (1 - t) * 0.75;
+}
