+++ src/ui/toast.js
@@ -1,0 +1,9 @@
+// toast.js
+// Simple toast notification helper.
+export function showToast(msg){
+  const t = document.getElementById('toast');
+  if (!t) return;
+  t.textContent = msg;
+  t.classList.add('show');
+  setTimeout(()=> t.classList.remove('show'), 2500);
+}
