+++ src/ui/tooltip.js
@@ -1,0 +1,4 @@
+// tooltip.js
+// Exports the floating tooltip element used for hover labels.
+import { tooltip } from '../core/scene.js';
+export { tooltip };
